Update(UT): re-org tests
