+++ apps/rules/src/tests/blockframe.spec.ts
@@ -2,51 +2,15 @@
   apps,
   assertFails,
   assertSucceeds,
-  initializeTestApp,
-  loadFirestoreRules,
-  initializeAdminApp,
 } from '@firebase/rules-unit-testing';
 import { testFixture } from './fixtures/data';
-import fs from 'fs';
-import { TokenOptions } from '@firebase/rules-unit-testing/dist/src/api';
-import { Movie } from '@blockframes/movie/+state';
-import { MovieAppAccess } from '@blockframes/utils/apps';
-import { StoreStatus, StoreType } from '@blockframes/utils/static-model';
+import { Firestore, initFirestoreApp } from '@blockframes/testing/firebase/functions';
 import { META_COLLECTION_NAME, MAINTENANCE_DOCUMENT_NAME } from '@blockframes/utils/maintenance';
 
 //Meta collection, for maintenance control.
 const metaDoc = `${META_COLLECTION_NAME}/${MAINTENANCE_DOCUMENT_NAME}`;
-
-type ExtractPromise<T> = T extends Promise<infer I> ? I : never;
-type PromiseFirestore = ReturnType<typeof initFirestoreApp>;
-type Firestore = ExtractPromise<PromiseFirestore>;
-
 testFixture[metaDoc] = {};
 testFixture[metaDoc].endedAt = true;
-
-async function initFirestoreApp(
-  projectId: string,
-  rulePath: string,
-  data: Record<string, Object> = {},
-  auth?: TokenOptions
-) {
-  //Define these env vars to avoid getting console warnings
-  process.env.GCLOUD_PROJECT = projectId;
-  process.env.FIRESTORE_EMULATOR_HOST = 'localhost:8080';
-  await setData(projectId, data);
-  const app = initializeTestApp({ projectId, auth });
-  await loadFirestoreRules({ projectId, rules: fs.readFileSync(rulePath, 'utf8') });
-
-  return app.firestore();
-}
-
-function setData(projectId: string, dataDB: Record<string, Object>) {
-  const app = initializeAdminApp({ projectId });
-  const db = app.firestore();
-  // Write data to firestore app
-  const promises = Object.entries(dataDB).map(([key, doc]) => db.doc(key).set(doc));
-  return Promise.all(promises);
-}
 
 describe('Blockframe In Maintenance', () => {
   const projectId = `rules-spec-${Date.now()}`;
@@ -106,78 +70,6 @@ describe('Blockframe Not In Maintenance', () => {
   test('Blockframe admin should be able to read all users', async () => {
     const usersRef = db.collection('users');
     await assertSucceeds(usersRef.get());
-  });
-});
-
-describe('Org Admin', () => {
-  const projectId = `rules-spec-${Date.now()}`;
-  let db: Firestore;
-
-  beforeAll(async () => {
-    db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, { uid: 'uid-admin' });
-  });
-
-  afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-  test('should allow org admin to read', async () => {
-    const usersRef = db.collection('users');
-    await assertSucceeds(usersRef.get());
-  });
-});
-
-describe('General User', () => {
-  const projectId = `rules-spec-${Date.now()}`;
-  let db: Firestore;
-
-  beforeAll(async () => {
-    db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, { uid: 'uid-user2' });
-  });
-
-  afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-  test('should not allow user to write /blockframesAdmin/user', async () => {
-    const usersRef = db.doc('blockframesAdmin/007');
-    await assertFails(usersRef.set({ uid: '007' }));
-  });
-});
-
-describe('Users Collection Rules Tests', () => {
-  const projectId = `rules-spec-${Date.now()}`;
-  const uid = 'uid-user2';
-  const user = { uid, email: 'u2@cascade8.com', note: 'JestTestSet-User2' };
-  let db: Firestore;
-
-  beforeAll(async () => {
-    db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, { uid: 'uid-user2' });
-  });
-
-  afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-  test('should allow user to read own user doc (own uid)', async () => {
-    const usersRef = db.doc('users/uid-user2');
-    await assertSucceeds(usersRef.get({}));
-  });
-
-  test('should allow user to delete own user doc (own uid)', async () => {
-    const usersRef = db.doc('users/uid-user2');
-    await assertSucceeds(usersRef.delete());
-  });
-
-  test("shouldn't allow user to create own user doc (own uid)", async () => {
-    const usersRef = db.doc('users/uid-user2');
-    await assertFails(usersRef.set({ email: 'u2@cascade8.com', note: 'JestTestSet-User2' }));
-  });
-
-  test('should allow user to create own user doc (own uid)', async () => {
-    const usersRef = db.doc(`users/${uid}`);
-    await assertSucceeds(usersRef.set(user));
-  });
-
-  test('should contain exact fields for own user docs', async () => {
-    const usersRef = db.doc(`users/${uid}`);
-    const expUserSnap = await usersRef.get();
-    const expUser = await expUserSnap.data();
-    expect(user).toEqual(expUser);
   });
 });
 
@@ -308,150 +200,6 @@ describe.skip('Permission Rules Tests', () => {
   });
 
   afterAll(() => Promise.all(apps().map((app) => app.delete())));
-});
-
-describe('Movies Rules Tests', () => {
-  const projectId = `rules-spec-${Date.now()}`;
-  let db: Firestore;
-
-  describe('With User in org', () => {
-    const newMovieId = 'MI-007';
-    const existMovieId = 'MI-077';
-    const storeConfig = {
-      status: <StoreStatus>'draft',
-      storeType: <StoreType>'Library',
-      appAccess: <MovieAppAccess>{ festival: true },
-    };
-
-    beforeAll(async () => {
-      db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, { uid: 'uid-user2' });
-    });
-
-    afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-    describe('Read Movie', () => {
-      test('should be able to read movie', async () => {
-        const movieRef = db.doc('movies/M001');
-        await assertSucceeds(movieRef.get());
-      });
-
-      test('should be able to read movie distribution rights', async () => {
-        const movieDRRef = db.doc('movies/M001/distributionRights/DR001');
-        await assertSucceeds(movieDRRef.get());
-      });
-    });
-
-    describe('Create Movie', () => {
-      test('storestatus other than draft should not be able', async () => {
-        const movieRef = db.doc(`movies/${newMovieId}`);
-        const createdMovie: Partial<Movie> = {
-          id: newMovieId,
-          storeConfig,
-        };
-        createdMovie.storeConfig['status'] = 'accepted';
-        await assertFails(movieRef.set(createdMovie));
-      });
-
-      test('without create permission for org should not be able', async () => {
-        const newMovieTitleUnavailableId = 'MI-000';
-        const movieRef = db.doc(`movies/${newMovieTitleUnavailableId}`);
-        const createdMovie: Partial<Movie> = {
-          id: newMovieTitleUnavailableId,
-          storeConfig,
-        };
-        await assertFails(movieRef.set(createdMovie));
-      });
-
-      test('with create permission for org, invalid id should not be able', async () => {
-        const newMovieTitleUnavailableId = 'MI-000';
-        const movieRef = db.doc(`movies/${newMovieId}`);
-        const createdMovie: Partial<Movie> = {
-          id: newMovieTitleUnavailableId,
-          storeConfig,
-        };
-        await assertFails(movieRef.set(createdMovie));
-      });
-
-      test('with create permission for org should be able', async () => {
-        const createdMovie: Partial<Movie> = {
-          id: newMovieId,
-          storeConfig: { status: <StoreStatus>'draft' } as any,
-        };
-        const movieDoc = db.collection('movies').doc(newMovieId).set(createdMovie);
-        await assertSucceeds(movieDoc);
-      });
-    });
-
-    describe('Update Movie', () => {
-      const fields: any = [
-        ['id', 'MI-0xx'],
-        ['_meta', { createdBy: '' }],
-        ['_meta', { createdAt: '' }],
-        ['_type', 'drama'],
-        ['storeConfig', { appAccess: { catalog: true } }],
-        ['storeConfig', { status: 'rejected' }],
-        ['storeConfig', { storeType: 'blah' }],
-        ['storeConfig', { appAccess: { festival: {} } }],
-      ];
-      test.each(fields)("updating restricted '%s' field shouldn't be able", async (key, value) => {
-        const movieRef = db.doc(`movies/${existMovieId}`);
-        const details = {};
-        details[key] = value;
-        await assertFails(movieRef.update(details));
-      });
-
-      test('user valid org, updating unrestricted field should be able', async () => {
-        const movieRef = db.doc(`movies/${existMovieId}`);
-        const movieDetailsOther = { notes: 'update in unit-test' };
-        await assertSucceeds(movieRef.update(movieDetailsOther));
-      });
-    });
-  });
-
-  describe('User with admin role', () => {
-    const draftMovieId = 'MI-0d7';
-
-    beforeAll(async () => {
-      db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, { uid: 'uid-admin' });
-    });
-
-    afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-    //TODO #4301
-    test('should not be able to delete movie', async () => {
-      const movieRef = db.doc(`movies/${draftMovieId}`);
-      await assertFails(movieRef.delete());
-    });
-  });
-
-  describe('With User not in org', () => {
-    const newMovieId = 'MI-007';
-    const draftMovieId = 'MI-0d7';
-    const newMovieDetails = { id: `${newMovieId}` };
-
-    beforeAll(async () => {
-      db = await initFirestoreApp(projectId, 'firestore.rules', testFixture, {
-        uid: 'uid-peeptom',
-      });
-    });
-
-    afterAll(() => Promise.all(apps().map((app) => app.delete())));
-
-    test("user without valid org shouldn't be able to read movie title", async () => {
-      const movieRef = db.doc('movies/M001');
-      await assertFails(movieRef.get());
-    });
-
-    test("user without valid org shouldn't be able to create movie title", async () => {
-      const movieRef = db.doc(`movies/${newMovieId}`);
-      await assertFails(movieRef.set(newMovieDetails));
-    });
-
-    test("user without valid org shouldn't be able to delete movie title", async () => {
-      const movieRef = db.doc(`movies/${draftMovieId}`);
-      await assertFails(movieRef.delete());
-    });
-  });
 });
 
 //TODO: 4200
