+++ libs/testing/src/lib/firebase/functions.ts
@@ -4,8 +4,11 @@ import { runChunks } from '@blockframes/firebase-utils';
 import { resolve } from 'path';
 import { config } from 'dotenv';
 import { firebase as firebaseEnv } from '@env';
+import { initializeTestApp, loadFirestoreRules, initializeAdminApp } from '@firebase/rules-unit-testing';
 import type { FeaturesList } from 'firebase-functions-test/lib/features';
 import type { AppOptions } from 'firebase-admin'; // * Correct Import
+import fs from 'fs';
+import { TokenOptions } from '@firebase/rules-unit-testing/dist/src/api';
 
 export interface FirebaseTestConfig extends FeaturesList {
   firebaseConfig?: { projectId: string }
@@ -45,6 +48,34 @@ export function initFunctionsTestMock(offline = true, overrideConfig?: AppOption
 export function getTestingProjectId() {
   // projectId cannot have '.' in the string; need whole numbers
   return 'test' + testIndex;
+}
+
+type ExtractPromise<T> = T extends Promise<infer I> ? I : never;
+type PromiseFirestore = ReturnType<typeof initFirestoreApp>;
+export type Firestore = ExtractPromise<PromiseFirestore>;
+
+export async function initFirestoreApp(
+  projectId: string,
+  rulePath: string,
+  data: Record<string, Object> = {},
+  auth?: TokenOptions
+) {
+  //Define these env vars to avoid getting console warnings
+  process.env.GCLOUD_PROJECT = projectId;
+  process.env.FIRESTORE_EMULATOR_HOST = 'localhost:8080';
+  await setData(projectId, data);
+  const app = initializeTestApp({ projectId, auth });
+  await loadFirestoreRules({ projectId, rules: fs.readFileSync(rulePath, 'utf8') });
+
+  return app.firestore();
+}
+
+function setData(projectId: string, dataDB: Record<string, Object>) {
+  const app = initializeAdminApp({ projectId });
+  const db = app.firestore();
+  // Write data to firestore app
+  const promises = Object.entries(dataDB).map(([key, doc]) => db.doc(key).set(doc));
+  return Promise.all(promises);
 }
